Move VertexInterpolatorTests.
