+++ src/test/java/net/imagej/ops/geom/DefaultVertexInterpolatorTest.java
@@ -27,12 +27,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric3d;
+package net.imagej.ops.geom;
 
 import static org.junit.Assert.assertEquals;
 
 import net.imagej.ops.AbstractOpTest;
 import net.imagej.ops.geom.BitTypeVertexInterpolator;
+import net.imagej.ops.geom.DefaultVertexInterpolator;
 
 import org.junit.Test;
 
@@ -42,41 +43,53 @@ import org.junit.Test;
  * @author Tim-Oliver Buchholz, University of Konstanz.
  *
  */
-public class BitTypeVertexInterpolatorTest extends AbstractOpTest {
+public class DefaultVertexInterpolatorTest extends AbstractOpTest {
 
 	@Test
-	public void interpolatorTest_v2() {
-			int[] p1 = new int[]{0,0,0};
-			int[] p2 = new int[]{10, 0, 10};
-			double v1 = 0;
-			double v2 = 1;
-			double[] res = (double[]) ops.run(BitTypeVertexInterpolator.class, p1, p2, v1, v2);
-				assertEquals(5, res[0], 1e-10);
-				assertEquals(0, res[1], 1e-10);
-				assertEquals(5, res[2], 1e-10);
+	public void vertexInterpolator_025_v2_Test() {
+		int[] p1 = new int[]{0,0,0};
+		int[] p2 = new int[]{10, 0, 10};
+		double v1 = 0;
+		double v2 = 1;
+		double[] res = (double[]) ops.run(DefaultVertexInterpolator.class, p1, p2, v1, v2, 0.25);
+		assertEquals(2.5, res[0], 1e-10);
+		assertEquals(0, res[1], 1e-10);
+		assertEquals(2.5, res[2], 1e-10);
 	}
 	
 	@Test
-	public void interpolatorTest_v1() {
-			int[] p1 = new int[]{0,0,0};
-			int[] p2 = new int[]{10, 0, 10};
-			double v1 = 1;
-			double v2 = 0;
-			double[] res = (double[]) ops.run(BitTypeVertexInterpolator.class, p1, p2, v1, v2);
-			assertEquals(5, res[0], 1e-10);
-			assertEquals(0, res[1], 1e-10);
-			assertEquals(5, res[2], 1e-10);
+	public void vertexInterpolator_1_v2_Test() {
+		int[] p1 = new int[]{0,0,0};
+		int[] p2 = new int[]{10, 0, 10};
+		double v1 = 0;
+		double v2 = 1;
+		double[] res = (double[]) ops.run(DefaultVertexInterpolator.class, p1, p2, v1, v2, 1);
+		assertEquals(10, res[0], 1e-10);
+		assertEquals(0, res[1], 1e-10);
+		assertEquals(10, res[2], 1e-10);
 	}
 	
 	@Test
-	public void interpolatorTest_equal() {
-			int[] p1 = new int[]{0,0,0};
-			int[] p2 = new int[]{10, 0, 10};
-			double v1 = 1;
-			double v2 = 1;
-			double[] res = (double[]) ops.run(BitTypeVertexInterpolator.class, p1, p2, v1, v2);
-			assertEquals(5, res[0], 1e-10);
+	public void vertexInterpolator_1_v1_Test() {
+		int[] p1 = new int[]{0,0,0};
+		int[] p2 = new int[]{10, 0, 10};
+		double v1 = 1;
+		double v2 = 0;
+		double[] res = (double[]) ops.run(DefaultVertexInterpolator.class, p1, p2, v1, v2, 1);
+			assertEquals(0, res[0], 1e-10);
 			assertEquals(0, res[1], 1e-10);
-			assertEquals(5, res[2], 1e-10);
+			assertEquals(0, res[2], 1e-10);
+	}
+	
+	@Test
+	public void vertexInterpolator_1_equalValues_Test() {
+		int[] p1 = new int[]{0,0,0};
+		int[] p2 = new int[]{10, 0, 10};
+		double v1 = 1;
+		double v2 = 1;
+		double[] res = (double[]) ops.run(DefaultVertexInterpolator.class, p1, p2, v1, v2, 1);
+		assertEquals(0, res[0], 1e-10);
+		assertEquals(0, res[1], 1e-10);
+		assertEquals(0, res[2], 1e-10);
 	}
 }
